implements search archive feature #238
