+++ client/routes/document.js
@@ -1,10 +1,9 @@
 var Templates = require('../templates');
 var initJqueryComp = require('../lib/init-jquery-components.js');
-var $ = require('jquery');
-var QueryString = require('querystring');
 var getData = require('../lib/get-data.js');
 var JSONToHTML = require('../../lib/transforms/json-to-html-data');
 var searchListener = require('../lib/search-listener');
+var archiveListeners = require('../lib/archive-listeners');
 
 module.exports = function (page) {
   page('/documents/:id', load, render, listeners);
@@ -40,43 +39,5 @@ function render (ctx, next) {
 function listeners (ctx, next) {
   initJqueryComp(ctx);
   searchListener();
-  $('.expand').on('submit', function (e) {
-    e.preventDefault();
-    var archiveTree = document.getElementById('archive-tree');
-    var documentID = this.id.slice(7);
-    var query = QueryString.parse(ctx.querystring.substr(1));
-    var url = ctx.pathname;
-
-    if (query.expanded) {
-      if (!Array.isArray(query.expanded)) {
-        query.expanded = [query.expanded];
-      }
-      var queryPos = query.expanded.indexOf(documentID);
-      if (queryPos > -1) {
-        query.expanded.splice(queryPos, 1);
-      } else {
-        query.expanded.push(documentID);
-      }
-    } else {
-      query.expanded = [];
-      query.expanded.push(documentID);
-    }
-
-    ctx.querystring = '?' + QueryString.stringify(query);
-
-    url += ctx.querystring;
-
-    ctx.path = url;
-
-    var opts = {
-      headers: { Accept: 'application/vnd.api+json' }
-    };
-
-    getData(url, opts, function (json) {
-      var data = JSONToHTML(json);
-      archiveTree.innerHTML = Templates['archiveTree'](data);
-      document.getElementsByTagName('title')[0].textContent = data.titlePage;
-      listeners(ctx, next);
-    });
-  });
+  archiveListeners(ctx, listeners);
 }
