+++ lib/transforms/json-to-html-data.js
@@ -15,10 +15,11 @@ module.exports = (resource) => {
   const date = getValues.getDate(resource.data);
   const links = getLinks(resource.data);
   const parent = getParent(resource);
+  const fonds = getFonds(resource);
   var identifier = getIdentifier(resource.data);
   identifier = identifier ? identifier.value : null;
 
-  return {title, titlePage, type, fact, related, description, details, displayLocation, system, level, date, links, identifier, parent};
+  return {title, titlePage, type, fact, related, description, details, displayLocation, system, level, date, links, identifier, parent, fonds};
 };
 
 function getType (data) {
@@ -335,6 +336,23 @@ function getParent (resource) {
   } else {
     return null;
   }
+}
+
+function getFonds (resource) {
+  var fondsID;
+  var fonds;
+  if (getNestedProperty(resource, 'data.relationships.fonds')) {
+    fondsID = resource.data.relationships.fonds.data[0].id;
+    fonds = resource.included.find(el => el.id === fondsID);
+    if (fonds) {
+      return {
+        id: fondsID,
+        link: fonds.links.self,
+        summary_title: fonds.attributes.summary_title
+      };
+    }
+  }
+  return null;
 }
 
 function getParentFact (resource) {
