+++ templates/partials/records/archive-tree.html
@@ -5,12 +5,16 @@
     </div>
     <div class="columns medium-4">
 
+    <form action="/search/documents" id="archive-search">
       <div class="searchbox searchbox--archive">
-        <input type="search" class="searchbox__search" placeholder="Or search within this archive">
-        <button aria-label="search" class="searchbox__submit">
+        <input type="search" class="searchbox__search" placeholder="Or search within this archive" name="q" id="archive-q">
+        <input type="hidden" name="archive" value="{{#if fonds}}{{fonds.summary_title}}{{else}}{{this.title}}{{/if}}" id="archive-title">
+        <button aria-label="search" type="submit" class="searchbox__submit">
           {{> global/icon i="search" size="24" }}
         </button>
       </div>
+    </form>
+
     </div>
 
     <div class="archive-tree columns">
